Support repositories groups and aliases
This change  adds the ability to specify objects in the .gitbulkconfig
file instead of just paths to repositories. Those objects allow users to
specify a short name or and a group for each repo. Those short names and
repos can be used for each command just like the paths can be used for
each command to specify which repos the action should be taken on.